+++ R/9-global.R
@@ -1,8 +1,6 @@
-#' @importFrom methods new as
+#' @importFrom methods new as show
 #' @importFrom plyr rbind.fill
 #' @importMethodsFrom kernlab predict
-#' @importClassesFrom Biobase ExpressionSet
-#' @importClassesFrom GEOquery GEODataTable
 #' @importMethodsFrom ROCR plot
 #' @importFrom graphics plot
 NULL